refactor(infra): разделена композиция AppDependencies
app composition разделена на focused extension-файлы для SwiftData bootstrap, runtime wiring, background refresh, URL cache и app-level actions
AppDependencies оставлен как контейнер собранных зависимостей, а bootstrap и runtime entry points вынесены ближе к своим зонам ответственности
