+++ README.md
@@ -519,7 +519,22 @@
 - [x] `Sort Unread Articles Contract Cleanup`: без legacy migration support переименовать persisted/settings contract сортировки из общего `sortMode` / `articleSortMode` в unread-specific contract, чтобы модель настроек, screen state и UI identifiers не сохраняли старую семантику общей сортировки статей;
 - [x] `Shared Loading Indicator Tint And Reuse`: расширить `AppRefreshIndicator` tint / foreground-style параметром и использовать общий custom refresh symbol для загрузки картинок в `CachedArticleImageView`; на экране проверки источника `Preview Feed` показывать тот же символ во вторичном цвете рядом с серым label, чтобы loading states не расходились визуально и не сохраняли системный `ProgressView` там, где уже есть app-owned refresh symbol;
 - [x] `Article Title Text Normalization`: нормализовать заголовок статьи до сохранения / показа в списке и reader: декодировать HTML entities вроде `&amp;#132;` / `&amp;#147;`, убрать невидимые управляющие символы из RSS / CDATA payload и сохранить читаемый display title без технических артефактов из feed XML;
-- [x] `Source Icon Appears On Add`: при сохранении нового или отредактированного источника сразу запускать загрузку / discovery иконки для этого feed, чтобы sidebar показывал source icon после добавления источника без ожидания ручного обновления;
+- [x] `Source Icon Appears On Add`: при сохранении нового или отредактированного источника сразу запускать загрузку / discovery иконки для этого feed, чтобы sidebar показывал source icon после добавления источника без ожидания ручного обновления.
+
+#### Post-Interface Cleanup
+- [x] `App Dependencies Composition Split`: разделить `AppDependencies` на явные composition-фабрики и runtime action surface: оставить в `AppDependencies` хранение собранных зависимостей, а bootstrap SwiftData / CloudKit, cache directories, background refresh wiring и app-level UI actions вынести в небольшие файлы / extension-группы с понятными границами ответственности;
+- [ ] `App Action Routing Cleanup`: отделить app-level navigation / Safari / source-management actions от infrastructure composition: `showInbox`, `selectArticle`, `openArticleInSafari`, source-management loaders и refresh entry points должны выглядеть как app action layer поверх `AppState` и services, а не как часть сборки dependency graph;
+- [ ] `Reader Opening Contract Rename`: без legacy migration support переименовать `ReaderMode` / `defaultReaderMode` в контракт открытия статьи, который прямо описывает выбор между app-owned feed reader и `SFSafariViewController`; убрать оставшуюся browser/reader-mode терминологию, которая была уместна до удаления `WKWebView` и full-text reader extension point;
+- [ ] `Unread Sort Contract Tightening`: сузить naming сортировки до unread-specific semantics на границе settings / screen state / query selection: убрать остаточную общую терминологию `sortMode` там, где она фактически означает настройку `Sort Unread Articles`, но не ломать repository-level сортировку, если она остаётся универсальным query parameter;
+- [ ] `Source Management State Partition`: разрезать `SourceManagementScreenState` на отдельные state/reducer файлы для `Add Feed`, `Move Source` и `Create Folder`, оставив root state только координатором presentation destination и derived view state; сохранить текущий behavior, но сделать каждый flow тестируемым и review-friendly отдельно;
+- [ ] `Source Management View Split`: разделить крупный `SourceManagementScreenView` на root navigation container, destination factory и отдельные view-файлы для add/edit feed, move source и create/edit folder, чтобы UI-слой не смешивал navigation plumbing, bindings и конкретную разметку форм;
+- [ ] `Reader Content Rendering Pipeline Split`: разнести `ArticleScreenContentRenderer` на доменные части: body block models, HTML block tokenizer / parser, inline text renderer, media URL resolver и final body assembly; renderer должен остаться thin orchestration layer между `ArticleScreenBodyPayloadNormalizer` и UI presentation blocks;
+- [ ] `Feed Parsing Module Split`: разделить `FeedParserService` на XML document builder, feed kind detection, RSS parser, Atom parser и parse pipeline diagnostics, чтобы будущие tests для parsing / normalizer не зависели от одного большого файла с несколькими уровнями ответственности;
+- [ ] `Refresh Pipeline Boundary Cleanup`: проверить `FeedRefreshService` после накопленных archive / retention / diagnostics изменений и выделить из него отдельные шаги conditional request, not-modified handling, fetched response reconciliation, diagnostics/log persistence и batch orchestration, не меняя публичный `FeedRefreshCoordinating` contract без необходимости;
+- [ ] `Loading Indicator Contract Audit`: пройти оставшиеся loading states после `Shared Loading Indicator Tint And Reuse` и явно зафиксировать, где допустим системный `ProgressView` / UIKit indicator, а где должен использоваться `AppRefreshIndicator`, чтобы UI не возвращался к смешанным indicator semantics перед UI tests;
+- [ ] `Test Harness Decomposition`: разложить `RSSReaderTestSupport` на focused builders / fixtures: SwiftData harness, scripted HTTP client, DTO factories, logging doubles и source-management fixtures; тесты должны импортировать общие helpers без неявной зависимости от feed refresh harness;
+- [ ] `App Dependencies Test Suite Split`: разделить большой `AppDependenciesTests` на suites по composition areas: SwiftData configuration, sync bootstrap policy, runtime orchestrator wiring, background refresh dependencies, cache directories и app action entry points; это должно подготовить инфраструктуру к эпику `Testing` и снизить стоимость targeted test runs;
+- [ ] `Roadmap TODO Cleanup`: перенести временные inline TODO для будущих OPML import/export flows из runtime UI-кода в roadmap / Post-MVP backlog или заменить их нейтральными disabled actions, чтобы текущий код не хранил незакрытые комментарии по отложенной функциональности.
 
 #### Testing
 - [ ] unit tests для normalizer;
